Major refactor
Significant changes across the board.

The beginning of v2 of QuickDash.

The purpose is to be able to include the package `@unicorn/quick-dash`
and begin a new dashboard project quickly and easily.

There is another project showing the boilerplate code to load the
project and configure.
