+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@unicorn/quick-dash",
-  "version": "1.0.0",
+  "version": "2.0.0-alpha",
   "description": "Quick Dashboards",
   "author": "Darryn Ten, Steve Ops",
   "private": true,
@@ -74,6 +74,7 @@
     "file-loader": "^2.0.0",
     "flush-promises": "^1.0.2",
     "friendly-errors-webpack-plugin": "^1.7.0",
+    "fs": "0.0.1-security",
     "html-webpack-plugin": "^3.2.0",
     "http-proxy-middleware": "^0.19.1",
     "inject-loader": "^4.0.1",
@@ -91,6 +92,7 @@
     "optimize-css-assets-webpack-plugin": "^5.0.1",
     "ora": "^3.0.0",
     "portfinder": "^1.0.20",
+    "require-dir": "^1.2.0",
     "rimraf": "^2.6.0",
     "sass-loader": "^7.1.0",
     "sass-resources-loader": "^2.0.0",
